resume-generator-api/requirements.txt
asgiref==3.6.0
Brotli==1.0.9
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
whitenoise==6.4.0
pdfkit==0.6.0